albaranes-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic>=0.40.0
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
pillow==11.0.0
pypdfium2==4.30.0
pydantic==2.9.2
